backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dotenv==0.9.9
fastapi==0.116.1
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.25.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
